Dev A Dev B
Also added the following:
try-catch to make sure input in amount of players is numeric.
adjusted to minimum amount of players to reflect the ruleset.
Updated classdiagram_v0.puml to reflect the code.
There was also a problem in the toString method being private as showed in the classdiagram_v0.puml.
Updated the setup checker for y/n which originally only checked for y so anything else would continue as n even if typed for example e.
Adjusted the code to use the new Player constructor which automatically adds 30,000 to the balance of a new player.
Deleted setBalance method as it was doing the same as the method deposit

Notes:
Should deposit not be change to "change  or update" so it could also be used for withdrawals.
FileIO and TextUI methods are and could be static. In that case the also don't need to be instantiated.